+++ gum-gum/gum-gum-htb.js
@@ -313,6 +313,12 @@ function GumGumHtb(configs) {
                 queryObj.pi = 3;
                 queryObj.si = parseInt(xSlot.inSlot, 10);
             }
+            if (xSlot.irisid) {
+                queryObj.irisid = xSlot.irisid;
+            }
+            if (xSlot.iriscat) {
+                queryObj.iriscat = xSlot.iriscat;
+            }
             queryObj.bf = xSlot.sizes.reduce(function (acc, curSlotDim) {
                 // Use the uriencoded version of ,
                 var deliminter = acc ? '%2C' : '';