+++ gum-gum/CHANGES.md
@@ -1,6 +1,6 @@
 
 # 2.3.0
-- Adds bF param
+- Adds bf, irisid, and iriscat params
 - Updates bidUnitInCents and removes multiplication of bidprice
 # 2.2.0
 - Adds support for CCPA (US Privacy)